brought over our nutrition api
swapped out our api's to be the nutritionx one since we swapped project ideas. current code is based around the idea of we make a first call to get the food id/name/brand(store) then from that the user can click on the food on our list to gain more information which our second call will use to take the food id and grab the addition information of the item ie. calories/sugar/protein... information.