made changes to the ajax call
the ajax call should now be able to tell the use when there are no page results(just need a not alert mandol?) then for the results in the search it should pull the item name, brand its associated with and the id of the food and append them into a list onto the html. We just need to link the html to the javascript and test them out to see if they break.