figured out how to get our second api call to work
the issue was that the .data had to only be set to food, since the elements data-food the .data already lookds for things with a data tag then the "food" specifies each one. so our api's second call now works all we need is the modal inside the html and we can test appenind the data inside of it.